embed files
Re-evaluating `util.HasPathPrefix`.
This was added to have a clear distinction between files and other arguments.
Quite often completion get overwhelming when a large amount gets listed between other arguments.
However, there is tag support now and while this is limited to zsh at the moment it aligns with future plans.
It has also been irritating for users not aware of this "feature" and _wrong_ in the sense of what can be passed to a command.