nix_completer: Add action for completing attributes on flakes
Fixes: https://github.com/carapace-sh/carapace-bin/issues/2374

This will only complete attributes for flakes that are in
the user's registry or are a local path. This means
that the flake eval cache should be very useful for these
completions, and that carapace won't go out and fetch
random flakes when the user is trying to tab complete.